Implemented Dependency Injection (DI) for entrypoint
