+++ src/Chirp.CLI/Chirp.CLI.csproj
@@ -12,6 +12,7 @@
     <ItemGroup>
       <PackageReference Include="CsvHelper" Version="30.0.1" />
       <PackageReference Include="docopt.net" Version="0.8.1" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0-rc.1.23419.4" />
     </ItemGroup>
 
 
@@ -19,6 +20,9 @@
     <ItemGroup>
       <ProjectReference Include="..\Chirp.SimpleDB\Chirp.SimpleDB.csproj" />
     </ItemGroup>
+
+
+
