+++ src/Chirp.CLI/Interfaces/IChirpHandler.cs
@@ -1,0 +1,6 @@
+namespace Chirp.CLI.Interfaces;
+
+public interface IChirpHandler
+{
+    void HandleInput(string[] args);
+}
