+++ src/Chirp.CLI/Program.cs
@@ -1,38 +1,21 @@
-using Chirp.CLI.Shared;
+using Chirp.CLI.Interfaces;
 using Chirp.CLI.Storage;
 using Chirp.CLI.Types;
 using Chirp.SimpleDB.Storage;
-using DocoptNet;
-
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Chirp.CLI;
 public static class Program
 {
-    private static IStorage<ChirpRecord>? _csvStorage;
     public static void Main(string[] args)
     {
-        var arguments = new Docopt().Apply(UserInterface.USAGE, args, exit: true);
-
-        _csvStorage = ChirpStorageProvider.Storage;
-
-        if (arguments["cheep"].IsTrue)
-        {
-            var author = Environment.UserName;
-            var message = args[1];
-            var timestamp = DateTimeHelper.DateTimeToEpoch(DateTime.Now);
-
-            var chirp = new ChirpRecord(author, message, timestamp); 
-
-            _csvStorage.StoreEntity(chirp);
-        }
-        else if(arguments["read"].IsTrue)
-        {
-            UserInterface.Read(_csvStorage.GetEntities());
-        }
-        else if (arguments["--help"].IsTrue)
-        {
-            UserInterface.Help();
-        }
+        var serviceProvider = new ServiceCollection()
+            .AddSingleton<IStorageProvider<ChirpRecord>, ChirpStorageProvider>()
+            .AddSingleton<IChirpHandler, ChirpHandler>()
+            .BuildServiceProvider();
+        
+        var handler = serviceProvider.GetService<IChirpHandler>();
+        handler.HandleInput(args);
     }
 }
